// src/app/shared/item-service/model/item-output.model.ts
export interface ItemOutputModel {
    codigoItem: number;
    nomenclaturaVarejo: string;
    nomeDetalhado: string;
    possuiItemAVencer: boolean;
    participaPbm: boolean;
    permiteAdesao: boolean;
    possuiKitAdesao: boolean;
    exclusivoPanvel: boolean;
    participaListaReferencial: boolean;
    participaFarmaciaPopular: boolean;
}